try again with unpackddelparam
There doesn't seem to be any direct way to tell if lparam is a pointer, ghandle or atom.  If it's an atom and bit 4 is not set unpackddelparam will try to use it as a pointer and crash with an uncatchable fault if (void *)atom doesn't happen to be valid.  So I try to access the address first to force a catchable fault.